Internal constants from int to short where makes sense

+++ src/main/java/org/threadly/concurrent/KeyDistributedExecutor.java
@@ -36,12 +36,12 @@ import org.threadly.util.ExceptionUtils;
  * @since 2.5.0 (existed since 1.0.0 as TaskExecutorDistributor)
  */
 public class KeyDistributedExecutor {
-  protected static final int DEFAULT_LOCK_PARALISM = 32;
+  protected static final short DEFAULT_LOCK_PARALISM = 32;
   protected static final float CONCURRENT_HASH_MAP_LOAD_FACTOR = 0.75f;  // 0.75 is ConcurrentHashMap default
-  protected static final int CONCURRENT_HASH_MAP_MIN_SIZE = 8;
-  protected static final int CONCURRENT_HASH_MAP_MAX_INITIAL_SIZE = 64;
-  protected static final int CONCURRENT_HASH_MAP_MAX_CONCURRENCY_LEVEL = 32;
-  protected static final int ARRAY_DEQUE_INITIAL_SIZE = 8;  // minimum is 8, should be 2^X
+  protected static final short CONCURRENT_HASH_MAP_MIN_SIZE = 8;
+  protected static final short CONCURRENT_HASH_MAP_MAX_INITIAL_SIZE = 64;
+  protected static final short CONCURRENT_HASH_MAP_MAX_CONCURRENCY_LEVEL = 32;
+  protected static final short ARRAY_DEQUE_INITIAL_SIZE = 8;  // minimum is 8, should be 2^X
   
   protected final Executor executor;
   protected final StripedLock sLock;
